atob() deprcated since v15.13.0/ v14.17.0
This function is only provided for compatibility with legacy web platform APIs and is advised to never be used in new code, because they use strings to represent binary data and predate the introduction of typed arrays in JavaScript. For code running using APIs, converting between base64-encoded strings and binary data states that it should be performed using Buffer.from(str, 'base64') and buf.toString('base64').